requirements.txt
google-generativeai==0.3.0
langchain==0.1.16
langchain-google-genai==0.0.13
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0